KVM: PPC: Book3S HV: Add config option to control real-mode timing code
This adds a new config option, CONFIG_KVM_BOOK3S_HV_EXIT_TIMING, which
controls whether or not we collect timing statistics for the real-mode
guest entry/exit code.  This is in line with the version of this code
submitted upstream.  This also converts sprintf to snprintf as requested
by upstream review.

Signed-off-by: Paul Mackerras <paulus@samba.org>