KVM: PPC: Make real_vmalloc_addr() public
This helper translates vmalloc'd addresses to linear addresses.
It is only used by KVM now and resides in the HV KVM code.
We will need it further in VFIO-SPAPR-TCE memory pre-registration code
which is going to reside in the powerpc code so let's make
real_vmalloc_addr() public and move it to the powerpc code as after all
it does not do anything special for KVM.

Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>