KVM: PPC: Book3S HV: Fix preempted vcore list locking and stolen time calculation
When a vcore gets preempted, we put it on the preempted vcore list for
the current CPU.  The runner task then calls schedule() and comes back
some time later and takes itself off the list.  We need to be careful
to lock the list that it was put onto, which may not be the list for the
current CPU since the runner task may have moved to another CPU.

Also, whenever the vcore state is VCORE_PREEMPT we need to be counting
stolen time for it.  This wasn't the case when we had a vcore that no
longer has any runnable threads in it but still has a runner task, so
we do an explicit call to kvmppc_core_start_stolen() in that case.

BZ: 123519, 123731

Signed-off-by: Paul Mackerras <paulus@samba.org>