powerpc/powernv: Add real mode version of xchg()
TCE tables should be invalidated in real mode using different
cache-inhibited assembler commands than in virtual mode.

This defines and implements xchg() callback for real mode.

The only user for it would be KVM and since it does not
use set/clear callbacks, there is no need in defining them
for real mode.

The exchange_rm callback is defined for all supported powernv platforms
(p5ioc2, ioda, ioda2).

This does not add flush_rm() as the only existing flush()
(for virtual mode) is defined for DART and we do not support KVM/VFIO
there.

Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>