powerpc/iommu: Add missing init for IOMMU initialization bus notifier
The DDW merge moved tce_iommu_bus_nb from the powernv platform code
to the common iommu code as it is done in the upstream now. However
machine_subsys_initcall_sync() was forgotten so the notifier was not
enabled and virtual functions were not getting assigned to groups.

This adds missing bus notifier initcall.

BZ: 124569

Signed-off-by: Wei Yang <weiyang@linux.vnet.ibm.com>
Signed-off-by: Alexey Kardashevskiy <aik@ozlabs.ru>
Signed-off-by: Paul Mackerras <paulus@samba.org>