crypto: change 842 alg to use software
Change the crypto 842 compression alg to use the software 842 compression
and decompression library.  Change the name of this crypto alg to "sw842".
Remove the fallback to LZO compression.

Previously, this crypto compression alg attemped 842 compression using
PowerPC hardware, and fell back to LZO compression and decompression if
the 842 PowerPC hardware was unavailable or failed.  This should not
fall back to any other compression method, however; users of this crypto
compression alg can fallback if desired, and transparent fallback tricks
callers into thinking they are getting 842 compression when they actually
get LZO compression - the failure of the 842 hardware should not be
transparent to the caller.

The crypto compression alg for a hardware device also should not be located
in crypto/ so this is now a software-only implementation that uses the 842
software compression/decompression library.  Since users of the "842" alg
expected hardware compression, the name of this software-only alg is
changed to "sw842"; the new hardware 842 crypto compression alg will be
aliased to "842" in a later patch.

Signed-off-by: Dan Streetman <ddstreet@ieee.org>
Signed-off-by: Dan Streetman <ddstreet@us.ibm.com>
Signed-off-by: Paul Mackerras <paulus@samba.org>