drivers/crypto/nx: add hardware 842 crypto comp alg
Add crypto compression alg for 842 hardware compression and decompression.

This crypto compression alg is named "nx842" to indicate it uses hardware
to perform the compression and decompression, while the software 842
compression alg is named "sw842".  However, since before this split there
was only one 842 compression alg named "842" which only used hardware,
this is also aliased "842" for backwards compatibility.

This uses only the PowerPC coprocessor hardware for 842 compression.  It
also uses the hardware for decompression, but if the hardware fails it will
fall back to the 842 software decompression library, so that decompression
never fails (for valid 842 compressed buffers).  A header must be used in
most cases, due to the hardware's restrictions on the buffers being
specifically aligned and sized.

Due to the header this driver adds, compressed buffers it creates cannot be
directly passed to the 842 software library for decompression.  However,
compressed buffers created by the software 842 library can be passed to
this driver for hardware 842 decompression (with the exception of buffers
containing the "short data" template, as lib/842/842.h explains).

Signed-off-by: Dan Streetman <ddstreet@ieee.org>
Signed-off-by: Dan Streetman <ddstreet@us.ibm.com>
Signed-off-by: Paul Mackerras <paulus@samba.org>