crypto: Fix 842 crypto name/driver_name
Bring the crypto/842.c and drivers/crypto/nx/nx-842-crypto.c drivers
up to the code that was accepted upstream, specifically by fixing
the crypto name to be "842" for both the generic and NX driver, and
using different crypto driver_names, "842-generic" for the sw only
implementation and "842-nx" for the NX implementation.

Add crypto driver priority to both drivers, 100 for the generic driver
and 300 for the NX driver.

Signed-off-by: Dan Streetman <ddstreet@us.ibm.com>
Signed-off-by: Paul Mackerras <paulus@samba.org>