KVM: PPC: Book3S HV: Send IPI to host core to wake VCPU
This patch adds support to real-mode KVM to search for a core
running in the host partition and send it an IPI message with
VCPU to be woken and the IPI handler code to wake up the VCPU.
This avoids having to switch to the host partition to complete
an H_IPI hypercall when the VCPU which is the target of the
the H_IPI is not loaded (is not running in the guest). The
patch also includes the support in the IPI handler running in
host to do the wakeup.

Signed-off-by: Suresh Warrier <warrier@linux.vnet.ibm.com>
Signed-off-by: Paul Mackerras <paulus@samba.org>