KVM: PPC: Book3S HV: Add tunable to disable H_IPI redirection
Although, redirecting the wakeup of a VCPU from the H_IPI
hypercall to a core running in the host is usually a good
idea, in one heavily interrpt-driven SMT1 workload, some
regression was observed. This patch adds a sysctl tunable
called h_ipi_redirect to control this feature. The default
value for this tunable is 1 - that is enable the feature.

Signed-off-by: Suresh Warrier <warrier@linux.vnet.ibm.com>
Signed-off-by: Paul Mackerras <paulus@samba.org>