KVM: PPC: Book3S: Correct width in XER handling
The Fixed Point Exception Register (XER) is defined in kvm_regs and
kvm_vcpu_arch as a 64 bit field on 64-bit machines and in most places
it is accessed as such.

This patch corrects several places where it is accessed as a 32 bit
field.  In some assembly code it is accessed via 32 bit load and store
instructions, which access the upper 32 bits in big-endian kernels.
In some C code it is cast to a u32, causing the upper 32 bits to be
cleared.  This could result in XER being cleared in the guest.

This patch corrects those places by extending the access methods to
64 bits.

Signed-off-by: Sam Bobroff <sam.bobroff@au1.ibm.com>
Signed-off-by: Paul Mackerras <paulus@samba.org>